Pull breakpoint helpers into own file
